requirements.txt
lxml>=4.9.0
pydantic>=2.0.0
email-validator>=2.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
jinja2>=3.1.0
python-multipart>=0.0.6
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-jose[cryptography]>=3.3.0
passlib==1.7.4
bcrypt==4.0.1
openpyxl>=3.0.0
reportlab>=4.0.0
matplotlib>=3.7.0
